c version 0.2.1

+++ c/src/channel.c
@@ -2,8 +2,9 @@
 // SPDX-FileCopyrightText: © 2024 HSE AG, <opensource@hseag.com>
 
 #include "channel.h"
+#include "evibase.h"
 
-Channel_t channel_init(double sample, double reference)
+Channel_t channel_init(uint32_t sample, uint32_t reference)
 {
     Channel_t ret = {.sample = sample, .reference = reference};
     return ret;
@@ -11,5 +12,5 @@ Channel_t channel_init(double sample, double reference)
 
 void channel_print(const Channel_t * self, FILE * stream, bool newLine)
 {
-    fprintf_s(stream, "Sample=%f, Reference=%f%s", self->sample, self->reference, newLine ? "\n" : "");
+    fprintf_s(stream, "Sample=%i, Reference=%i%s", self->sample, self->reference, newLine ? "\n" : "");
 }
